Partial work on ring point generation.
While working on this commit, I realized that I'd like to modify the
center-pt-finding and the first-ring-pt-finding simultaneously so that it's more
clear that everything means what I want it to, and that things line up
mathematically as we want them to. The only code seems to work but I think is a
tiny bit hacky.