bark.lua/stick bark: fix choice of top ring point.
Basically, good triangles in a polyhedral column will be acute triangles, which
means, for our rings, that the first top ring point needs to be looking-down-
clockwise from the bottom's first ring point. I don't have a proof of this
statement about acute triangles, but it feels intuitively correct and looks
good, so I'm making an informed guess that it's correct.