Work on a replacement leaf glob algorithm.
Instead of starting from scratch, I've decided to try to save the iterative
approach by being more careful about how the triangulation is recalculated each
time we add a new point. To be honest, I'm not 100% this will work, but at least
it recognizes that adding a new point to the convex hull is not as simple as
replacing a single old triangle with 3 new ones.