Add line-drawing support to VertexArray.
The out directions of branching points are not quite acting as I'd expect. Next
up is to debug that. Then try a bit more to get this very simple canopy symmetry
approach to look a bit better. If that doesn't work, I can move onto a slightly
more sophisticated approach.